+++ packages/perseus-linter/src/rules/all-rules.ts
@@ -7,7 +7,9 @@
 import AbsoluteUrl from "./absolute-url";
 import DoubleSpacingAfterTerminal from "./double-spacing-after-terminal";
 import ExpressionWidget from "./expression-widget";
+import ExpressionWidgetError from "./expression-widget-error";
 import ExtraContentSpacing from "./extra-content-spacing";
+import FreeResponseWidgetError from "./free-response-widget-error";
 import HeadingLevel1 from "./heading-level-1";
 import HeadingLevelSkip from "./heading-level-skip";
 import HeadingSentenceCase from "./heading-sentence-case";
@@ -19,8 +21,11 @@ import ImageSpacesAroundUrls from "./image-spaces-around-urls";
 import ImageUrlEmpty from "./image-url-empty";
 import ImageWidget from "./image-widget";
 import InaccessibleWidget from "./inaccessible-widget";
+import InteractiveGraphWidgetError from "./interactive-graph-widget-error";
+import LabelImageWidgetError from "./label-image-widget-error";
 import LinkClickHere from "./link-click-here";
 import LongParagraph from "./long-paragraph";
+import MatcherWidgetError from "./matcher-widget-error";
 import MathAdjacent from "./math-adjacent";
 import MathAlignExtraBreak from "./math-align-extra-break";
 import MathAlignLinebreaks from "./math-align-linebreaks";
@@ -31,6 +36,10 @@ import MathStartsWithSpace from "./math-starts-with-space";
 import MathTextEmpty from "./math-text-empty";
 import MathWithoutDollars from "./math-without-dollars";
 import NestedLists from "./nested-lists";
+import NumericInputWidgetError from "./numeric-input-widget-error";
+import PhetSimulationWidgetError from "./phet-simulation-widget-error";
+import PythonProgramWidgetError from "./python-program-widget-error";
+import RadioWidgetError from "./radio-widget-error";
 import StaticWidgetInQuestionStem from "./static-widget-in-question-stem";
 import TableMissingCells from "./table-missing-cells";
 import UnbalancedCodeDelimiters from "./unbalanced-code-delimiters";
@@ -70,4 +79,13 @@ export default [
     ImageSpacesAroundUrls,
     ImageWidget,
     InaccessibleWidget,
+    RadioWidgetError,
+    ExpressionWidgetError,
+    FreeResponseWidgetError,
+    MatcherWidgetError,
+    NumericInputWidgetError,
+    PhetSimulationWidgetError,
+    PythonProgramWidgetError,
+    LabelImageWidgetError,
+    InteractiveGraphWidgetError,
 ];