+++ packages/perseus/src/widget-container.tsx
@@ -176,6 +176,9 @@ class WidgetContainer extends React.Component<Props, State> {
                                         widgetType: type,
                                         widgetId: this.props.id,
                                         message: error.message,
+                                        stack:
+                                            error.stack ??
+                                            "No stack trace available",
                                         userAgent: userAgent,
                                     },
                                 });