+++ packages/perseus/src/widgets/image/__docs__/image-demo.stories.tsx
@@ -19,7 +19,12 @@ import {
 import {
     earthMoonImage,
     frescoImage,
-    monasteryImage,
+    graphieImage,
+    graphieImageAlt,
+    portraitImage,
+    portraitImageCaption,
+    portraitImageLongDescription,
+    portraitImageTitle,
     scienceImage,
     scienceImageAlt,
     scienceImageCaption,
@@ -72,12 +77,10 @@ export const ImageMobile: Story = {
 export const ImagePortrait: Story = {
     decorators: [imageRendererDecorator],
     args: {
-        backgroundImage: monasteryImage,
-        caption:
-            "Kalenić Monastery, after 1407, Serbia (photo: [Ванилица](https://commons.wikimedia.org/wiki/File:Wiki_%C5%A0umadija_XI_Kaleni%C4%87_Monastery_874.jpg), CC BY-SA 4.0)",
-        title: "Kalenić Monastery",
-        longDescription:
-            "Later architecture in Serbia, notably that of the so-called Morava School, is smaller and more decorative, often utilizing the so-called Athonite plan (with choroi and subsidiary chapels), as at Ravanica (1370s), with five domes, or the smaller and simpler Kalenić (after 1407).",
+        backgroundImage: portraitImage,
+        caption: portraitImageCaption,
+        title: portraitImageTitle,
+        longDescription: portraitImageLongDescription,
     },
 };
 
@@ -140,6 +143,14 @@ export const DecorativeImageMobile: Story = {
         longDescription:
             "This is a *very* long description of the earth and moon.",
         decorative: true,
+    },
+};
+
+export const GraphieImage: Story = {
+    decorators: [imageRendererDecorator],
+    args: {
+        backgroundImage: graphieImage,
+        alt: graphieImageAlt,
     },
 };